a space
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@13234 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ShotStatistics.cxx
@@ -75,7 +75,7 @@ void ShotStatistics::recordFire(FlagType* flag, const float *pVec, const float *
   double cos = acos(dot);
   double radToDeg = 180.0/3.1415;
 
-  lastShotDeviation = cos*radToDeg;
+  lastShotDeviation = (float)(cos*radToDeg);
 }
 
 void ShotStatistics::recordHit(FlagType* flag)
